refactor: separate classes into own files
