+++ src/dgt.power.common/IXrmConnectionFactory.cs
@@ -1,0 +1,10 @@
+// Copyright (c) DIGITALL Nature. All rights reserved
+// DIGITALL Nature licenses this file to you under the Microsoft Public License.
+
+namespace dgt.power.common;
+
+public interface IXrmConnectionFactory
+{
+    IXrmConnection GetDefault();
+    IXrmConnection GetWithProfile(string profile);
+}
