+++ src/modules/dgt.power.analyzer/Logic/ActiveLayerAnalyze.cs
@@ -5,13 +5,12 @@ using System.Diagnostics;
 using dgt.power.analyzer.Base;
 using dgt.power.analyzer.Reports;
 using dgt.power.common;
-using dgt.power.common.Logic;
 using dgt.power.dataverse;
 using dgt.power.dto;
 using Microsoft.Xrm.Sdk;
 using Microsoft.Xrm.Sdk.Messages;
 using Microsoft.Xrm.Sdk.Metadata;
- using Spectre.Console;
+using Spectre.Console;
 
 namespace dgt.power.analyzer.Logic;
