+++ src/modules/dgt.power.analyzer/Logic/EntityAllAssetsAnalyze.cs
@@ -7,7 +7,6 @@ using dgt.power.analyzer.Base;
 using dgt.power.analyzer.Base.Config;
 using dgt.power.analyzer.Reports;
 using dgt.power.common;
-using dgt.power.common.Logic;
 using dgt.power.dataverse;
 using dgt.power.dto;
 using Microsoft.Xrm.Sdk.Messages;
