+++ src/modules/dgt.power.analyzer/Logic/NoActiveLayerAnalyze.cs
@@ -5,7 +5,6 @@ using System.Diagnostics;
 using dgt.power.analyzer.Base;
 using dgt.power.analyzer.Reports;
 using dgt.power.common;
-using dgt.power.common.Logic;
 using dgt.power.dataverse;
 using dgt.power.dto;
 using Microsoft.Xrm.Sdk.Messages;
