+++ src/modules/dgt.power.analyzer/Logic/RedundantPatchAnalyze.cs
@@ -5,7 +5,6 @@ using System.Globalization;
 using dgt.power.analyzer.Base;
 using dgt.power.analyzer.Reports;
 using dgt.power.common;
-using dgt.power.common.Logic;
 using dgt.power.dataverse;
 using dgt.power.dto;
 using Spectre.Console;
