+++ src/modules/dgt.power.export/Base/BaseExport.cs
@@ -6,7 +6,6 @@ using System.Text.Json.Serialization;
 using System.Text.RegularExpressions;
 using dgt.power.common;
 using dgt.power.common.FileAccess;
-using dgt.power.common.Logic;
 
 namespace dgt.power.export.Base;
