+++ src/modules/dgt.power.export/Logic/CalendarExport.cs
@@ -4,7 +4,6 @@
 using System.Diagnostics;
 using dgt.power.common;
 using dgt.power.common.FileAccess;
-using dgt.power.common.Logic;
 using dgt.power.dataverse;
 using dgt.power.export.Base;
 using Microsoft.Xrm.Sdk;
