+++ src/modules/dgt.power.import/Base/BaseImport.cs
@@ -3,7 +3,6 @@
 
 using System.Diagnostics;
 using dgt.power.common;
-using dgt.power.common.Logic;
 using dgt.power.dto;
 
 namespace dgt.power.import.Base;
