+++ src/modules/dgt.power.import/Logic/BulkDeleteImport.cs
@@ -5,7 +5,6 @@ using System.Diagnostics;
 using System.Globalization;
 using dgt.power.common;
 using dgt.power.common.Extensions;
-using dgt.power.common.Logic;
 using dgt.power.dataverse;
 using dgt.power.dto;
 using dgt.power.import.Base;
