+++ src/modules/dgt.power.import/Logic/OutlookTemplateImport.cs
@@ -5,7 +5,6 @@ using System.Diagnostics;
 using System.Xml.Linq;
 using dgt.power.common;
 using dgt.power.common.Extensions;
-using dgt.power.common.Logic;
 using dgt.power.dataverse;
 using dgt.power.dto;
 using dgt.power.import.Base;
