+++ src/modules/dgt.power.import/Logic/SecureConfigImport.cs
@@ -3,7 +3,6 @@
 
 using System.Diagnostics;
 using dgt.power.common;
-using dgt.power.common.Logic;
 using dgt.power.dataverse;
 using dgt.power.dto;
 using dgt.power.import.Base;
