+++ src/modules/dgt.power.import/Logic/UserRoleImport.cs
@@ -4,7 +4,6 @@
 using System.Diagnostics;
 using dgt.power.common;
 using dgt.power.common.Extensions;
-using dgt.power.common.Logic;
 using dgt.power.dataverse;
 using dgt.power.dto;
 using dgt.power.import.Base;
