+++ src/modules/dgt.power.maintenance/Logic/FilterPowerFxPluginSteps.cs
@@ -3,7 +3,6 @@
 
 using System.Diagnostics;
 using dgt.power.common;
-using dgt.power.common.Logic;
 using dgt.power.dataverse;
 using dgt.power.maintenance.Base;
 using dgt.power.maintenance.Base.Config;
