+++ src/modules/dgt.power.push/PushCommand.cs
@@ -5,7 +5,6 @@ using System.Globalization;
 using System.Reflection;
 using System.Runtime.InteropServices;
 using dgt.power.common;
-using dgt.power.common.Logic;
 using dgt.power.dataverse;
 using dgt.power.push.Base;
 using dgt.power.push.Logic;
