+++ tests/dgt.power.tests/CommandTestContextBuilder.cs
@@ -4,7 +4,6 @@
 using System.Reflection;
 using System.Runtime.CompilerServices;
 using dgt.power.common;
-using dgt.power.common.Logic;
 using dgt.power.dataverse;
 using dgt.power.tests.Extensions;
 using FakeXrmEasy;
